rename elementsJets.* to jets.*
No need for the longer name within the elements directory.
