+++ Haskell/cbits/elements/jets.c
@@ -1,4 +1,4 @@
-#include "primitive/elements/elementsJets.h"
+#include "primitive/elements/jets.h"
 #include "../wrappers.h"
 
 WRAP_(version)
